requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.2
openai==1.73.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-docx==1.1.2
python-dotenv==1.1.0
reportlab==4.3.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
whitenoise==6.9.0
django-import-export==4.3.7
